+++ lib/browser/Browser.tsx
@@ -45,7 +45,7 @@
  * - `ui-browser-toolbar` - placed on the root `<div>` around the toolbar control.
  *
  * #### Properties
- * - `home: {string} ('about:blank')` - The site visited when the user clicks on the
+ * - `home=about:blank {string}` - The site visited when the user clicks on the
  * "home" button in the toolbar.  If this is empty, then 'about:blank' is used.
  * - `kbBack="alt+left" {string}` - Moves to the prior page if there is one.
  * - `kbForward="alt+right" {string}` - Moves to the next page if there is one.
@@ -54,12 +54,12 @@
  * - `kbPreviousSearch="alt+f3" {string}` - Go to the prior search term
  * - `kbRefresh="ctrl+r" {string}` - Reload the current webpage
  * - `kbSnapshot="" {string}` - Takes a snapshot of the current page.
- * - `notooltips: {boolean} (false)` - When set to true the tooltips will be suppresed.
+ * - `notooltips=false {boolean}` - When set to true the tooltips will be suppresed.
  * They are shown by default.
- * - `uri: {string} ('about:blank')` - The site the user selects to visit when the control
+ * - `uri=about:blank {string}` - The site the user selects to visit when the control
  * is created. If this is empty, then the home directory is used.  If the home directory
  * is empty, then the site is set to 'about:blank'
- * - `useparser: {boolean} (false)` - If true, then the onClip() event will take the
+ * - `useparser=false {boolean}` - If true, then the onClip() event will take the
  * HTML string and parse it into its DOM elements.  By default this is false because it
  * is a performance hit to parse it.
  *
@@ -67,6 +67,8 @@
  */
 
 const debug = require("debug")("gadgets.Browser");
+const debugCreate = require("debug")("gadgets.Browser:create");
+const debugRender = require("debug")("gadgets.Browser:render");
 
 import autobind from "autobind-decorator";
 import {WebviewTag} from "electron";
@@ -227,6 +229,8 @@ export class Browser extends BaseComponent<BrowserProps, BrowserState> {
 			uri: uri,
 			uriHistory: List(uri)
 		};
+
+		debugCreate("props: %O, state: %O", this.props, this.state);
 	}
 
 	/**
@@ -458,6 +462,8 @@ export class Browser extends BaseComponent<BrowserProps, BrowserState> {
 
 	public render() {
 		this.updateClassName();
+
+		debugRender("props: %O, state: %O", this.props, this.state);
 
 		return (
 			<Wrapper {...this.props}>